+++ scripts/custom_tools.py
@@ -4,6 +4,19 @@ from typing import List
 from openai import pydantic_function_tool
 from pydantic import BaseModel
 
+class ShouldExtractKeywords(BaseModel):
+    """A tool that initiates keyword entity extraction.
+
+    Attributes
+    ----------
+    use_tool : bool
+        Should the ExtractKeywordEntities tool be used.
+
+    """
+
+    use_tool: bool
+
+
 class ExtractKeywordEntities(BaseModel):
     """
     A model for extracting keyword entities.
@@ -31,6 +44,6 @@ class ExportDataToTSV(BaseModel):
     export: bool
 
 toolbox = [
-    pydantic_function_tool(ExtractKeywordEntities),
-    pydantic_function_tool(ExportDataToTSV)
-    ]
+    pydantic_function_tool(ShouldExtractKeywords),
+    pydantic_function_tool(ExportDataToTSV),
+]